Mixed

+++ Assets/Scripts/HealthSystem.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 using Unity.Entities;
 
 public partial class HealthSystem : SystemBase
@@ -11,7 +12,18 @@ public partial class HealthSystem : SystemBase
         {
             if(healthComponent.currentHealth <= 0)
             {
-                EntityManager.DestroyEntity(entity);
+                if(HasComponent<PlayerComponent>(entity))
+                {
+                    Debug.Log("Lost");
+                    SceneManager.LoadScene("Menu");
+                }
+                else
+                {
+                    EntityManager.DestroyEntity(entity);
+                    var enemyCount = PlayerPrefs.GetInt("enemyCount") - 1;
+                    PlayerPrefs.SetInt("enemyCount", enemyCount);
+                }
+                
             }
         }).Run();
     }
